Emesary MP fixe
A bug in the MP bridge will cause notification loss when a bridge is receiving more than one notification type; this is fixed in 2019.2 - however the most compatible fix for us is to split the in flight and hit notifications over two bridges. Possibly this is a better way of doing things because it allows more fine tuning of the bridge size.