F-14 APC rewrite
This is a new (Approach Power Compensator) based on the
actual aircraft data that I finally found.
"NASA TM-X-81833, Figure 38 Schematic of autothrottle control law".

This is a much more accurate and realistic than the previous APC which
has now been removed.

Details of the model specific to this F-14: https://i.imgur.com/gAbAGeE.png