F-14 FCS Use simulated sensors for angular velocities
There appears to be a time related issue that is contributing
towards the jitter - because the sensors on the aircraft feedback
the current measured value rather than knowing what it is.

After tuning the delay of 0.007 appears to give good results without
undue delay or overshoot.

It is because the delay affects the delay or overshoot that I think
this is probably the correct way to fix this.