F-14: Rework external stores select animation. Fixes issue #149
Instead of the current system using string[5] we have a single bool for each 3d model element.

There is a property rules file that sets the nicely named bools (which are aliased to the multiplay bools).

The 3d model only uses sim/multiplay/generic/bool[..] because this is the most efficient and avoids needing to have Nasal running in the model.