F-15: Fix target lock over MP
The target lock was never cleared for a target that was on the ground or various other conditions because the clearing needs to be done earlier. I've mode the clearing of the target to much earlier on in the loop and it seems OK although possibly it might get cleared and reset in different frames with a lot of targets.

Also fixed the VSD so that it clears the target line (w3_22) when the target is off display.