F-14: rework windshield lights.
New textures with better colouring and probably a more accurate font; although I couldn't find any indication of the typeface in the aircraft so I chose something that is close to what is likely. Not sure about the colours used so opted for red and green.

The SAM/AI/AA lights now have lettering and are in the right place.