F-14: Fix the hard step on the main struts.
Main problem was that the new gears (12,13) were being considered as fuselage strikes when in fact they should be ignored. We model the main gear as two struts because that is the best match for the aircraft where each main landing gear shock strut consists of an upper outer cylinder and a lower internal piston that has a maximum stroke of 25 inches. A hard step (31,000 pounds required for further compression) in the strut air curve provides a consistent 4-inch stroke remaining in the ground static condition.

Tested and seems to work fine at both the normal touchdown (660 ft/min) and the design max (~1500ft/min)