F-14 Rework of exterior lights (MP/normal)
Change all of the exterior lights from being Nasal controlled to be controlled entirely by XML animations including the flash/steady (using an expression based on sim time).

This gives much better results when viewing over MP because the flash will be controlled by the local sim-time and thus it will actually happen on time whereas before it relied on the properties being transmitted over MP which wasn't reliable.

Also implemented the POS lights properly so that the tail and wing are individually selectable.